requirements.txt
Sanic==18.12.0
Jinja2==2.10
python-socketio==3.1.2
aiofiles==0.4.0
alembic==1.0.7
httptools==0.0.11
multidict==4.4.2
psycopg2==2.7.7
python-editor==1.0.3
SQLAlchemy==1.2.8